requirements.txt
pyserial
paho-mqtt
numpy